Add files via upload
The entire program will be 3 parts:
Part 1: The conversion from .sra into .fastq format: DCCR.
Part 2: The conversion from .fastq into .bam.
Part 3: The Analysis of the files.

Part 2: THE QIACIA, or "Qualitycontrol, Indexing, Aligning, Conversion, Indexing, And other" script is a script that uses FastQC, bwa-mem, and samtools to automate the process of converting the files from fastq into aligned .bam files. It first needs to be Quality controlled using FastQC, it also dumps the report in a directory named after the sample. Then it is the indexing using the bwa-index function to the Human Genome (GRCh38). Alinging is then done through bwa-mem, conversion from .sam into .bam is done with samtools: view and sort. Second indexing is done with samtools index.