Update DCCR
This is the updated DCCR, DCCR 2.0 if you will. It comes with a checkpoint when removing the files. This way, if there is a fault within the script, it locates into which program that is faulty as well as does not remove the intermediary files. However, only if the program is faulty. If the program runs as it should, it will always remove the intermediary files. Unless one wants to specify it and comment out the "Removal" part.